+++ src/app/thrift-services/damsel/merchant-statistics.service.ts
@@ -3,7 +3,7 @@ import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 
 import { KeycloakTokenInfoService } from '../../keycloak-token-info.service';
-import { ThriftService } from '../thrift-service';
+import { ThriftService } from '../services/thrift/thrift-service';
 import { StatRequest, StatResponse } from './gen-model/merch_stat';
 import * as MerchantStatistics from './gen-nodejs/MerchantStatistics';
 import { StatRequest as ThriftStatRequest } from './gen-nodejs/merch_stat_types';
