+++ src/app/thrift-services/deanonimus/deanonimus.service.ts
@@ -2,7 +2,7 @@ import { Injectable, NgZone } from '@angular/core';
 import { Observable } from 'rxjs';
 
 import { KeycloakTokenInfoService } from '../../keycloak-token-info.service';
-import { ThriftService } from '../thrift-service';
+import { ThriftService } from '../services/thrift/thrift-service';
 import { SearchHit } from './gen-model/deanonimus';
 import * as Deanonimus from './gen-nodejs/Deanonimus';
