+++ src/app/thrift-services/utils/create-next-id.ts
@@ -1,0 +1,3 @@
+export function createNextId(ids: number[]) {
+    return Math.max(...ids) + 1;
+}
